requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-RESTful==0.3.9
grpcio==1.39.0
grpcio-tools==1.39.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
protobuf==3.17.3
pycodestyle==2.7.0
pytz==2021.1
requests==2.26.0
six==1.15.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0